requirements_base.txt
alabaster==0.7.10
appdirs==1.4.3
arrow==0.5.0
asgi-redis==0.14.1
asgiref==0.14.0
attrs==16.3.0
autobahn==0.16.0
Automat==0.5.0
awscli==1.11.56
Babel==2.3.4
beautifulsoup4==4.3.2
bibtexparser==0.6.0
boto==2.42.0
botocore==1.5.19
channels==0.17.3
colorama==0.3.7
constantly==15.1.0
contextlib2==0.5.4
coverage==4.2
daphne==0.14.3
dj-database-url==0.4.1
Django==1.8.8
django-countries==4.0
django-easymoney==0.7.1
django-extensions==1.5.5
django-floppyforms==1.7.0
django-idmap==0.4.1
django-jquery==3.1.0
django-jquery-ui==1.11.4.1
django-matrix-field==0.2
django-mysql==1.1.0
django-radiogrid==1.0.3
django-vanilla-views==1.0.4
djangorestframework==3.4.6
docutils==0.13.1
dropbox==2.2.0
future==0.14.3
honcho==0.7.1
huey==1.2.0
imagesize==0.7.1
incremental==16.10.1
IPy==0.83
Jinja2==2.9.5
jmespath==0.9.1
lxml==3.7.3
MarkupSafe==1.0
memoized-property==1.0.2
mendeley==0.3.2
mock==2.0.0
msgpack-python==0.4.8
oauthlib==0.7.2
otree-core==1.2.7
otree-save-the-change==1.1.3
packaging==16.8
pbr==1.10.0
Pillow==2.6.2
py==1.4.31
pyasn1==0.2.3
Pygments==2.2.0
pyparsing==2.2.0
pytest==2.9.2
pytest-django==3.0.0
python-dateutil==2.6.0
python-decouple==2.3
python-docx==0.8.5
python-redis-lock==3.2.0
pytz==2016.6.1
PyYAML==3.12
raven==5.25.0
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.4.2
rsa==3.4.2
s3transfer==0.1.10
schema==0.6.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
Twisted==16.2.0
txaio==2.5.1
unicodecsv==0.14.1
Unipath==1.1
urllib3==1.20
whitenoise==3.2.1
ws4py==0.3.5
XlsxWriter==0.9.3
xlwt==1.1.2
zope.interface==4.2.0
